refactor(opencode): extract the opencode-family shared core
Phase 1 of PR A (plan-opencode-family-shared-integration.md §10) —
behavior-preserving, opencode-only:

- agents/opencode-family.js: explicit family registry (single member) +
  shared FAMILY_EVENT_MAP/FAMILY_CAPABILITIES + isOpencodeFamily[Entry]
- hooks/opencode-family-plugin/core.mjs: the plugin runtime as a
  4-param factory (agentId/hookSource/logFileName/sessionIdPrefix);
  per-process state lives in the factory closure, one instance per
  entry-module evaluation
- hooks/opencode-family-plugin/session-ids.mjs: prefix-dependent helpers
  (normalize/resolve/isChild/cleanup + DEFAULT_SESSION_ID) built by
  createSessionIdHelpers(prefix); event extractors stay module-level
- hooks/opencode-plugin/index.mjs: thin entry (default export only, #413)
- hooks/opencode-family-install.js + thin opencode-install.js wrapper:
  full legacy surface preserved (named exports, return shapes incl. the
  opencode-not-found reason, CLI entry); registered plugin path is
  byte-identical (same hooks/ __dirname base)
- tests: session suite migrated to the family module; new family-core
  suite (factory isolation, dual-prefix matrix incl. the cross-prefix
  miss that motivated the sessionIdPrefix param, registry↔entry literal
  drift lock, #413 for every member entry)

Suite: 5230 tests, 0 fail.
