+++ test/opencode-family-core.test.js
@@ -1,0 +1,216 @@
+// opencode-family shared core — factory isolation, prefix matrix, and
+// registry↔entry cross-checks (plan-opencode-family-shared-integration.md §9).
+
+const { describe, it } = require("node:test");
+const assert = require("node:assert");
+const fs = require("fs");
+const path = require("path");
+const { pathToFileURL } = require("node:url");
+
+const {
+  OPENCODE_FAMILY,
+  FAMILY_EVENT_MAP,
+  FAMILY_CAPABILITIES,
+  isOpencodeFamily,
+  isOpencodeFamilyEntry,
+} = require("../agents/opencode-family");
+
+const HOOKS_DIR = path.join(__dirname, "..", "hooks");
+
+async function loadCore() {
+  const modulePath = path.join(HOOKS_DIR, "opencode-family-plugin", "core.mjs");
+  return import(pathToFileURL(modulePath).href);
+}
+
+async function loadSessionIds() {
+  const modulePath = path.join(HOOKS_DIR, "opencode-family-plugin", "session-ids.mjs");
+  return import(pathToFileURL(modulePath).href);
+}
+
+const OPENCODE_PARAMS = Object.freeze({
+  agentId: "opencode",
+  hookSource: "opencode-plugin",
+  logFileName: "opencode-plugin.log",
+  sessionIdPrefix: "opencode:",
+});
+// mimocode lands with the #607 rebase; the factory must already support any
+// second member — use its future params to prove instance isolation today.
+const MIMOCODE_PARAMS = Object.freeze({
+  agentId: "mimocode",
+  hookSource: "mimocode-plugin",
+  logFileName: "mimocode-plugin.log",
+  sessionIdPrefix: "mimocode:",
+});
+
+describe("opencode-family plugin factory", () => {
+  it("requires all four identity params", async () => {
+    const { createOpencodeFamilyPlugin } = await loadCore();
+    assert.throws(() => createOpencodeFamilyPlugin(), /agentId is required/);
+    for (const missing of ["agentId", "hookSource", "logFileName", "sessionIdPrefix"]) {
+      const params = { ...OPENCODE_PARAMS };
+      delete params[missing];
+      assert.throws(
+        () => createOpencodeFamilyPlugin(params),
+        new RegExp(`${missing} is required`),
+        `expected missing ${missing} to throw`
+      );
+    }
+  });
+
+  it("keeps two instances fully isolated (state maps + prefixes)", async () => {
+    const { createOpencodeFamilyPlugin } = await loadCore();
+    const oc = createOpencodeFamilyPlugin(OPENCODE_PARAMS);
+    const mc = createOpencodeFamilyPlugin(MIMOCODE_PARAMS);
+
+    // Separate mutable state: the parent maps are distinct objects.
+    assert.notStrictEqual(oc.__test._sessionParentById, mc.__test._sessionParentById);
+
+    // Same raw session id yields per-agent namespaced ids — the #607-review
+    // collision scenario (state.js keys sessions by session_id alone).
+    const ocBody = oc.__test.buildStateBody("idle", "SessionStart", "ses_123");
+    const mcBody = mc.__test.buildStateBody("idle", "SessionStart", "ses_123");
+    assert.strictEqual(ocBody.session_id, "opencode:ses_123");
+    assert.strictEqual(mcBody.session_id, "mimocode:ses_123");
+    assert.strictEqual(ocBody.agent_id, "opencode");
+    assert.strictEqual(mcBody.agent_id, "mimocode");
+    assert.strictEqual(ocBody.hook_source, "opencode-plugin");
+    assert.strictEqual(mcBody.hook_source, "mimocode-plugin");
+
+    // Child bookkeeping in one instance never leaks into the other.
+    oc.__test._sessionParentById.set("opencode:ses_child", "opencode:ses_root");
+    assert.strictEqual(
+      oc.__test.buildStateBody("working", "PreToolUse", "ses_child").headless,
+      true
+    );
+    assert.strictEqual(
+      mc.__test.buildStateBody("working", "PreToolUse", "ses_child").headless,
+      undefined
+    );
+    const mcIdle = mc.__test.translateEvent({ type: "session.idle", properties: { sessionID: "ses_child" } });
+    assert.strictEqual(mcIdle.event, "Stop"); // not SessionEnd — no cross-instance child state
+    oc.__test._sessionParentById.clear();
+  });
+});
+
+describe("opencode-family session-id helpers (prefix matrix)", () => {
+  for (const prefix of ["opencode:", "mimocode:"]) {
+    it(`${prefix} raw + prefixed child lookup, deleted removes one, disposed clears all`, async () => {
+      const { createSessionIdHelpers } = await loadSessionIds();
+      const ids = createSessionIdHelpers(prefix);
+      const map = new Map();
+      map.set(`${prefix}ses_child1`, `${prefix}ses_root`);
+      map.set(`${prefix}ses_child2`, `${prefix}ses_root`);
+
+      // raw and already-prefixed forms both hit
+      assert.strictEqual(ids.isChildSessionId("ses_child1", map), true);
+      assert.strictEqual(ids.isChildSessionId(`${prefix}ses_child1`, map), true);
+      assert.strictEqual(ids.isChildSessionId("ses_root", map), false);
+
+      // session.deleted removes exactly its own entry
+      ids.cleanupSessionParentMap(
+        { type: "session.deleted", properties: { sessionID: "ses_child1" } },
+        map
+      );
+      assert.strictEqual(map.has(`${prefix}ses_child1`), false);
+      assert.strictEqual(map.has(`${prefix}ses_child2`), true);
+
+      // server.instance.disposed clears everything, even without a sessionID
+      ids.cleanupSessionParentMap({ type: "server.instance.disposed", properties: {} }, map);
+      assert.strictEqual(map.size, 0);
+    });
+  }
+
+  it("helpers from one prefix never match another prefix's map keys", async () => {
+    const { createSessionIdHelpers } = await loadSessionIds();
+    const oc = createSessionIdHelpers("opencode:");
+    const mimoMap = new Map([["mimocode:ses_child", "mimocode:ses_root"]]);
+
+    // The v3-review blocker scenario: an opencode-prefixed lookup against a
+    // mimocode-keyed map must MISS — proving these helpers are prefix-bound
+    // and must come from the factory, never shared verbatim.
+    assert.strictEqual(oc.isChildSessionId("ses_child", mimoMap), false);
+    oc.cleanupSessionParentMap(
+      { type: "session.deleted", properties: { sessionID: "ses_child" } },
+      mimoMap
+    );
+    assert.strictEqual(mimoMap.size, 1); // wrong-prefix delete is a no-op
+  });
+
+  it("DEFAULT_SESSION_ID and resolve fallback follow the prefix", async () => {
+    const { createSessionIdHelpers } = await loadSessionIds();
+    const mc = createSessionIdHelpers("mimocode:");
+    assert.strictEqual(mc.DEFAULT_SESSION_ID, "mimocode:default");
+    assert.strictEqual(mc.resolveSessionId(null, null), "mimocode:default");
+    assert.strictEqual(mc.resolveSessionId("ses_a", null), "mimocode:ses_a");
+  });
+});
+
+describe("opencode-family registry", () => {
+  it("membership is the explicit allowlist, never eventSource inference", () => {
+    assert.strictEqual(isOpencodeFamily("opencode"), true);
+    // plugin-event agents that are NOT family members (plan §7):
+    assert.strictEqual(isOpencodeFamily("openclaw"), false);
+    assert.strictEqual(isOpencodeFamily("hermes"), false);
+    assert.strictEqual(isOpencodeFamily("claude-code"), false);
+    assert.strictEqual(isOpencodeFamily(null), false);
+  });
+
+  it("isOpencodeFamilyEntry keys off the PUBLIC agentId field", () => {
+    assert.strictEqual(isOpencodeFamilyEntry({ agentId: "opencode" }), true);
+    assert.strictEqual(isOpencodeFamilyEntry({ agentId: "claude-code" }), false);
+    assert.strictEqual(isOpencodeFamilyEntry({ familyAgentId: "opencode" }), false);
+    assert.strictEqual(isOpencodeFamilyEntry(null), false);
+  });
+
+  it("agents/opencode.js sources the shared family contract", () => {
+    const opencodeAgent = require("../agents/opencode");
+    assert.strictEqual(opencodeAgent.eventMap, FAMILY_EVENT_MAP);
+    assert.strictEqual(opencodeAgent.capabilities, FAMILY_CAPABILITIES);
+    assert.strictEqual(opencodeAgent.eventSource, "plugin-event");
+  });
+
+  it("every member's plugin entry literals match the registry (no drift)", async () => {
+    // The Bun-side entries cannot require this CJS registry, so they repeat
+    // the four identity params as literals — this test is the drift lock
+    // (plan §3.1 CJS/ESM note).
+    for (const [agentId, cfg] of Object.entries(OPENCODE_FAMILY)) {
+      const entryPath = path.join(HOOKS_DIR, cfg.pluginDirName, "index.mjs");
+      if (!fs.existsSync(entryPath)) {
+        // mimocode's entry lands with the #607 rebase; opencode must exist NOW.
+        assert.notStrictEqual(agentId, "opencode", "opencode entry must exist");
+        continue;
+      }
+      const source = fs.readFileSync(entryPath, "utf8");
+      const expectations = {
+        agentId,
+        hookSource: cfg.hookSource,
+        logFileName: cfg.logFileName,
+        sessionIdPrefix: cfg.sessionIdPrefix,
+      };
+      for (const [key, expected] of Object.entries(expectations)) {
+        const m = source.match(new RegExp(`${key}:\\s*"([^"]+)"`));
+        assert.ok(m, `${entryPath} must set ${key} as a string literal`);
+        assert.strictEqual(m[1], expected, `${entryPath} ${key} drifted from the registry`);
+      }
+
+      // #413: the entry module must have exactly one export — default.
+      const mod = await import(pathToFileURL(entryPath).href);
+      assert.deepStrictEqual(Object.keys(mod), ["default"]);
+      assert.strictEqual(typeof mod.default, "function");
+    }
+  });
+
+  it("registry config paths match the installer defaults", () => {
+    const opencodeInstall = require("../hooks/opencode-install");
+    const cfg = OPENCODE_FAMILY.opencode;
+    const os = require("os");
+    assert.strictEqual(
+      opencodeInstall.DEFAULT_PARENT_DIR,
+      path.join(os.homedir(), ...cfg.configDirSegments)
+    );
+    assert.strictEqual(
+      opencodeInstall.DEFAULT_CONFIG_PATH,
+      path.join(os.homedir(), ...cfg.configDirSegments, cfg.configFileName)
+    );
+  });
+});
